Move LogGuard to core logging
